+++ warshall.c
@@ -6,7 +6,7 @@
 
 void transitive_closure(unsigned *R, int n)
 {
-    register size_t rowsize;
+    register int rowsize;
     register unsigned mask;
     register unsigned *rowj;
     register unsigned *rp;
@@ -57,7 +57,7 @@ void transitive_closure(unsigned *R, int n)
 
 void reflexive_transitive_closure(unsigned *R, int n)
 {
-    register size_t rowsize;
+    register int rowsize;
     register unsigned mask;
     register unsigned *rp;
     register unsigned *relend;
